organized code and made a .gitingo
