+++ code/mood.py
@@ -1,14 +1,15 @@
-import datetime
 import random
-import speech_recognition as sr
 import pyttsx3
 
 engine = pyttsx3.init("sapi5")
 voices = engine.getProperty("voices")
 engine.setProperty("voice", voices[0].id)
 
-#list of words
-happy_list = ["happy", "great", "good", "better", "joyful", "content", "fulfilled", "positive", "optimistic", "upbeat", "elated", "radiant", "grateful", "blissful", "satisfied", "cheerful", "exuberant", "vibrant", "ecstatic", "peaceful", "serene", "blessed", "delighted", "glorious", "heartened", "inspired", "triumphant", "zealous", "merry", "euphoric", "jovial", "gleeful", "festive", "lively", "spirited", "buoyant", "uplifting", "hopeful", "comforted", "empowered"]
+def speak(audio):
+    engine.say(audio)
+    engine.runAndWait()
+#list of word that trigger the responses 
+happy_list = ["happy", "great", "good", "better", "joyful", "content", "fulfilled", "positive", "optimistic", "upbeat", "elated", "radiant", "grateful", "blissful", "satisfied", "cheerful", "exuberant", "vibrant", "ecstatic", "peaceful", "serene", "blessed", "delighted", "glorious", "heartened", "inspired", "triumphant", "zealous", "merry", "euphoric", "jovial", "gleeful", "festive", "lively", "spirited", "buoyant", "uplifting", "hopeful", "comforted", "empowered", "wonderful"]
 
 sad_list = ["sad", "negative", "boring", "melancholy", "unhappy", "disheartened", "discouraged", "gloomy", "dismal", "despondent", "downcast", "mournful", "blue", "downhearted", "crestfallen", "dejected", "forlorn", "sullen", "somber", "lugubrious", "woeful", "heartbroken", "bereaved", "mournful", "pensive", "doleful", "woebegone", "disconsolate", "lugubrious", "wistful", "regretful", "downtrodden", "defeated", "low-spirited", "demoralized", "dispirited", "apathetic", "listless", "weary", "dreary", "stifled", "glum", "uninspired", "unmotivated", "unfulfilled"]
 
@@ -17,55 +18,7 @@ angry_list = ["angry", "irate", "enraged", "furious", "livid", "incensed", "outr
 depressed_list = ["depressed", "downhearted", "melancholic", "disheartened", "blue", "low", "despondent", "despairing", "dismal", "gloomy", "forlorn", "sorrowful", "mournful", "wretched", "hopeless", "discouraged", "downtrodden", "disconsolate", "downcast", "miserable", "heavy-hearted", "sad", "unhappy", "glum", "joyless", "lugubrious", "morose", "somber", "sullen", "woeful", "tearful", "weepy", "weeping", "sulky", "pessimistic", "defeated", "crestfallen", "brokenhearted", "dejected", "desolate", "heartbroken", "inconsolable", "morbid", "unconsolable"]
 
 
-def speak(audio):
-    engine.say(audio)
-    engine.runAndWait()
-
-def wishme():
-    hour = int(datetime.datetime.now().hour)
-    if 0 <= hour < 12:
-        print("Good morning!")
-        speak("Good morning!")
-    elif 12 <= hour < 18:
-        print("Good afternoon")
-        speak("Good afternoon!")
-    else:
-        print("Good evening!")
-        speak("Good evening!")
-
-wishme()
-
-def speak_and_print(message):
-    print(message)
-    speak(message)
-
-speak_and_print("Before we start today, please state your name for easy communication")
-name = input("Your Name: ")
-print(name)
-
-
-print(f"I am Thea, your therapist for today, {name}. Let's start. How are you feeling today?")
-speak(f"I am Thea, your therapist for today, {name}. Let's start. How are you feeling today?")
-
-
-def takecommand():
-    # It takes microphone input and returns string output
-    r = sr.Recognizer()
-    with sr.Microphone() as source:
-        print("Listening...")
-        r.pause_threshold = 1
-        audio = r.listen(source)
-
-    try:
-        print("Recognizing...")
-        user_statement = r.recognize_google(audio, language="en-us")
-        print(f"User side: {user_statement}\n")
-
-    except Exception as e:
-        print("Say that again, please...")
-        return "None"
-    return user_statement
-
+# def for mood specificy responses 
 def happy_mood_responses():
     responses = [
         "Tell me more. It sounds interesting.",
@@ -160,26 +113,3 @@ def depressed_mood_responses():
     response = random.choice(responses)
     print(f"Thea: {response}")
     speak(response)
-
-
-
-if __name__ == "__main__":
-    
-    while True:
-        user_statement = takecommand().lower()
-
-        if any(word in user_statement for word in happy_list):
-            happy_mood_responses()
-
-        if any(word in user_statement for word in sad_list):
-            sad_mood_responses()
-
-        elif any(word in user_statement for word in angry_list):
-            angry_mood_responses()
-
-        elif any(word in user_statement for word in depressed_list):
-            depressed_mood_responses()
-
-        else:
-            print("I can't answer you. I'm sorry.")
-            speak("I can't answer you. I'm sorry.")
